Initial commit.
Super quick hack to get a gif "decoded".

This is ugly code and largely special cased for a single test input. The
basic idea is all here though.

Decoding will decode a header and ext blocks upon initialization. The
data contained in those blocks is held in memory. The frames are then
lazily decoded with each call to `get_frame`. No internal references to
the frames are kept after parsing so for each `get_frame` call ruby is
free to gc the `Frame` object.

The binary stream is also read into memory in chunks. It is entirely
possible I've done that wrong and I'm actually storing all the data in
memory. I'll figure that out later...

Next steps here are:
* test what is already written
* clean up the code
* parse the ext blocks (as of now they are just wrappers around unparsed
binary blobs)
* extend the decoding process to handle clearing the code table
* get real gif test cases that are more complex